Committing the functionalities for Searching

+++ IR-01_Maven/LuceneInformationRetrival/src/org/lir/model/Model.java
@@ -5,43 +5,51 @@ public class Model {
 	private String title;
 	private String summary;
 	private int rank;
-	private int relScore;
+	private float relScore;
 	private String path;
+
 	public String getTitle() {
 		return title;
 	}
+
 	public void setTitle(String title) {
 		this.title = title;
 	}
+
 	public String getSummary() {
 		return summary;
 	}
+
 	public void setSummary(String summary) {
 		this.summary = summary;
 	}
+
 	public int getRank() {
 		return rank;
 	}
+
 	public void setRank(int rank) {
 		this.rank = rank;
 	}
-	public int getRelScore() {
+
+	public float getRelScore() {
 		return relScore;
 	}
-	public void setRelScore(int relScore) {
+
+	public void setRelScore(float relScore) {
 		this.relScore = relScore;
 	}
+
 	public String getPath() {
 		return path;
 	}
+
 	public void setPath(String path) {
 		this.path = path;
 	}
-	
+
 	@Override
 	public String toString() {
 		return "Model [title=" + title + ", summary=" + summary + ", rank=" + rank + ", relScore=" + relScore + "]";
 	}
-	
-	
 }
